Divided into flavours.
