+++ Converter/app/premium/release/output-metadata.json
@@ -1,0 +1,20 @@
+{
+  "version": 1,
+  "artifactType": {
+    "type": "APK",
+    "kind": "Directory"
+  },
+  "applicationId": "com.converter.premium",
+  "variantName": "premiumRelease",
+  "elements": [
+    {
+      "type": "SINGLE",
+      "filters": [],
+      "properties": [],
+      "versionCode": 16,
+      "versionName": "1.4.3-premium",
+      "enabled": true,
+      "outputFile": "app-premium-release.apk"
+    }
+  ]
+}
