+++ Converter/app/stock/release/output-metadata.json
@@ -1,0 +1,20 @@
+{
+  "version": 1,
+  "artifactType": {
+    "type": "APK",
+    "kind": "Directory"
+  },
+  "applicationId": "com.converter.stock",
+  "variantName": "stockRelease",
+  "elements": [
+    {
+      "type": "SINGLE",
+      "filters": [],
+      "properties": [],
+      "versionCode": 16,
+      "versionName": "1.4.3-stock",
+      "enabled": true,
+      "outputFile": "app-stock-release.apk"
+    }
+  ]
+}
